+++ api/app.js
@@ -2,12 +2,12 @@ const createError = require("http-errors");
 const express = require("express");
 const path = require("path");
 const logger = require("morgan");
-const JWT = require("jsonwebtoken");
 
 const postsRouter = require("./routes/posts");
 const commentsRouter = require("./routes/comments");
 const authenticationRouter = require("./routes/authentication");
 const usersRouter = require("./routes/users");
+const tokenChecker = require('./lib/tokenChecker')
 
 const app = express();
 
@@ -17,26 +17,6 @@ app.use(express.json());
 app.use(logger("dev"));
 app.use(express.json());
 app.use(express.static(path.join(__dirname, "public")));
-
-// middleware function to check for valid tokens
-const tokenChecker = (req, res, next) => {
-  let token;
-  const authHeader = req.get("Authorization");
-
-  if (authHeader) {
-    token = authHeader.slice(7);
-  }
-
-  JWT.verify(token, process.env.JWT_SECRET, (err, payload) => {
-    if (err) {
-      console.log(err);
-      res.status(401).json({ message: "auth error" });
-    } else {
-      req.user_id = payload.user_id;
-      next();
-    }
-  });
-};
 
 // route setup
 app.use("/posts", tokenChecker, postsRouter);
